:truck: move delivery.js from pages to components
